Update `tests` structure
