+++ tests/geometry/costs_test.py
