+++ tests/geometry/pointcloud_test.py
