+++ tests/initializers/linear/sinkhorn_init_test.py
@@ -20,14 +20,10 @@ import numpy as np
 import pytest
 
 import ott.initializers.nn.initializers
-from ott.geometry import geometry, low_rank, pointcloud
+from ott.geometry import geometry, pointcloud
 from ott.initializers.linear import initializers as lin_init
-from ott.initializers.linear import initializers_lr
-from ott.initializers.quadratic import initializers as quad_init
 from ott.problems.linear import linear_problem
-from ott.problems.quadratic import quadratic_problem
-from ott.solvers.linear import sinkhorn, sinkhorn_lr
-from ott.solvers.quadratic import gromov_wasserstein
+from ott.solvers.linear import sinkhorn
 
 
 def create_sorting_problem(rng, n, epsilon=0.01, online=False):
@@ -333,256 +329,3 @@ class TestSinkhornInitializers:
     # check initializer is better
     if lse_mode:
       assert base_num_iter >= meta_num_iter
-
-
-class TestLRInitializers:
-
-  @pytest.mark.fast.with_args("kind", ["pc", "lrc", "geom"], only_fast=0)
-  def test_create_default_initializer(self, rng: jnp.ndarray, kind: str):
-    n, d, rank = 110, 2, 3
-    x = jax.random.normal(rng, (n, d))
-    geom = pointcloud.PointCloud(x)
-
-    if kind == "pc":
-      pass
-    elif kind == "lrc":
-      geom = geom.to_LRCGeometry()
-      assert isinstance(geom, low_rank.LRCGeometry)
-    elif kind == "geom":
-      geom = geometry.Geometry(geom.cost_matrix)
-    else:
-      raise NotImplementedError(geom)
-    prob = linear_problem.LinearProblem(geom)
-
-    solver = sinkhorn_lr.LRSinkhorn(rank=rank, initializer=None)
-    initializer = solver.create_initializer(prob)
-
-    assert initializer.rank == rank
-    if kind in ("pc", "lrc"):
-      assert isinstance(initializer, initializers_lr.KMeansInitializer)
-    else:
-      assert isinstance(initializer, initializers_lr.RandomInitializer)
-
-    q, r, g = initializer(prob)
-
-    assert q.shape == (n, rank)
-    assert r.shape == (n, rank)
-    assert g.shape == (rank,)
-
-  def test_explicitly_passing_initializer(self):
-    rank = 2
-    initializer = initializers_lr.RandomInitializer(rank=rank)
-    solver = sinkhorn_lr.LRSinkhorn(rank=rank, initializer=initializer)
-
-    assert solver.create_initializer(prob="not used") is initializer
-
-  @pytest.mark.parametrize(
-      "initializer", ["random", "rank2", "k-means", "generalized-k-means"]
-  )
-  @pytest.mark.parametrize("partial_init", ["q", "r", "g"])
-  def test_partial_initialization(
-      self, rng: jnp.ndarray, initializer: str, partial_init: str
-  ):
-    n, d, rank = 100, 10, 6
-    key1, key2, key3, key4 = jax.random.split(rng, 4)
-    x = jax.random.normal(key1, (n, d))
-    pc = pointcloud.PointCloud(x, epsilon=5e-1)
-    prob = linear_problem.LinearProblem(pc)
-    q_init = jax.random.normal(key2, (n, rank))
-    r_init = jax.random.normal(key2, (n, rank))
-    g_init = jax.random.normal(key2, (rank,))
-
-    solver = sinkhorn_lr.LRSinkhorn(rank=rank, initializer=initializer)
-    initializer = solver.create_initializer(prob)
-
-    if partial_init == "q":
-      q, _, _ = initializer(prob, q=q_init)
-      np.testing.assert_array_equal(q, q_init)
-    elif partial_init == "r":
-      _, r, _ = initializer(prob, r=r_init)
-      np.testing.assert_array_equal(r, r_init)
-    elif partial_init == "g":
-      _, _, g = initializer(prob, g=g_init)
-      np.testing.assert_array_equal(g, g_init)
-    else:
-      raise NotImplementedError(partial_init)
-
-  @pytest.mark.fast.with_args("rank", [2, 4, 10, 13], only_fast=True)
-  def test_generalized_k_means_has_correct_rank(
-      self, rng: jnp.ndarray, rank: int
-  ):
-    n, d = 100, 10
-    x = jax.random.normal(rng, (n, d))
-    pc = pointcloud.PointCloud(x, epsilon=5e-1)
-    prob = linear_problem.LinearProblem(pc)
-
-    solver = sinkhorn_lr.LRSinkhorn(
-        rank=rank, initializer="generalized-k-means"
-    )
-    initializer = solver.create_initializer(prob)
-
-    q, r, g = initializer(prob)
-
-    assert jnp.linalg.matrix_rank(q) == rank
-    assert jnp.linalg.matrix_rank(r) == rank
-
-  def test_generalized_k_means_matches_k_means(self, rng: jnp.ndarray):
-    n, d, rank = 120, 15, 5
-    eps = 1e-1
-    key1, key2 = jax.random.split(rng, 2)
-    x = jax.random.normal(key1, (n, d))
-    y = jax.random.normal(key1, (n, d))
-
-    pc = pointcloud.PointCloud(x, y, epsilon=eps)
-    geom = geometry.Geometry(cost_matrix=pc.cost_matrix, epsilon=eps)
-    pc_problem = linear_problem.LinearProblem(pc)
-    geom_problem = linear_problem.LinearProblem(geom)
-
-    solver = sinkhorn_lr.LRSinkhorn(
-        rank=rank, initializer="k-means", max_iterations=5000
-    )
-    pc_out = solver(pc_problem)
-
-    solver = sinkhorn_lr.LRSinkhorn(
-        rank=rank, initializer="generalized-k-means", max_iterations=5000
-    )
-    geom_out = solver(geom_problem)
-
-    with pytest.raises(AssertionError):
-      np.testing.assert_allclose(pc_out.costs, geom_out.costs)
-
-    np.testing.assert_allclose(
-        pc_out.reg_ot_cost, geom_out.reg_ot_cost, atol=0.5, rtol=0.02
-    )
-
-  @pytest.mark.parametrize("epsilon", [0., 1e-1])
-  def test_better_initialization_helps(self, rng: jnp.ndarray, epsilon: float):
-    n, d, rank = 81, 13, 3
-    key1, key2 = jax.random.split(rng, 2)
-    x = jax.random.normal(key1, (n, d))
-    y = jax.random.normal(key2, (n, d))
-    pc = pointcloud.PointCloud(x, y, epsilon=5e-1)
-    prob = linear_problem.LinearProblem(pc)
-
-    solver_random = sinkhorn_lr.LRSinkhorn(
-        rank=rank, epsilon=epsilon, initializer="random", max_iterations=10000
-    )
-    solver_init = sinkhorn_lr.LRSinkhorn(
-        rank=rank, epsilon=epsilon, initializer="k-means", max_iterations=10000
-    )
-
-    out_random = solver_random(prob)
-    out_init = solver_init(prob)
-
-    assert out_random.converged
-    assert out_init.converged
-    # converged earlier
-    assert (out_init.errors > -1).sum() < (out_random.errors > -1).sum()
-    # converged to a better solution
-    assert out_init.reg_ot_cost < out_random.reg_ot_cost
-
-
-class TestQuadraticInitializers:
-
-  @pytest.mark.parametrize("kind", ["pc", "lrc", "geom"])
-  def test_create_default_lr_initializer(self, rng: jnp.ndarray, kind: str):
-    n, d1, d2, rank = 150, 2, 3, 5
-    eps = 1e-1
-    key1, key2 = jax.random.split(rng, 2)
-    x = jax.random.normal(key1, (n, d1))
-    y = jax.random.normal(key1, (n, d2))
-    kwargs_init = {"foo": "bar"}
-
-    geom_x = pointcloud.PointCloud(x, epsilon=eps)
-    geom_y = pointcloud.PointCloud(y, epsilon=eps)
-    if kind == "pc":
-      pass
-    elif kind == "lrc":
-      geom_x = geom_x.to_LRCGeometry()
-      geom_y = geom_y.to_LRCGeometry()
-    elif kind == "geom":
-      geom_x = geometry.Geometry(geom_x.cost_matrix, epsilon=eps)
-      geom_y = geometry.Geometry(geom_y.cost_matrix, epsilon=eps)
-    else:
-      raise NotImplementedError(kind)
-    prob = quadratic_problem.QuadraticProblem(geom_x, geom_y)
-
-    solver = gromov_wasserstein.GromovWasserstein(
-        rank=rank, quad_initializer=None, kwargs_init=kwargs_init
-    )
-    initializer = solver.create_initializer(prob)
-
-    assert isinstance(initializer, quad_init.LRQuadraticInitializer)
-    assert initializer.rank == rank
-    linear_init = initializer._linear_lr_initializer
-    if kind in ("pc", "lrc"):
-      assert isinstance(linear_init, initializers_lr.KMeansInitializer)
-    else:
-      assert isinstance(linear_init, initializers_lr.RandomInitializer)
-    assert linear_init._kwargs == kwargs_init
-
-  def test_non_lr_initializer(self):
-    solver = gromov_wasserstein.GromovWasserstein(
-        rank=-1, quad_initializer="not used"
-    )
-    initializer = solver.create_initializer(prob="not used")
-    assert isinstance(initializer, quad_init.QuadraticInitializer)
-
-  @pytest.mark.parametrize("rank", [-1, 2])
-  def test_explicitly_passing_initializer(self, rank: int):
-    if rank == -1:
-      linear_init = lin_init.SortingInitializer()
-      q_init = quad_init.QuadraticInitializer()
-    else:
-      linear_init = initializers_lr.Rank2Initializer(rank)
-      q_init = quad_init.LRQuadraticInitializer(linear_init)
-
-    solver = gromov_wasserstein.GromovWasserstein(
-        initializer=linear_init,
-        quad_initializer=q_init,
-    )
-
-    assert solver.linear_ot_solver.initializer is linear_init
-    assert solver.quad_initializer is q_init
-    if solver.is_low_rank:
-      assert solver.quad_initializer.rank == rank
-
-  @pytest.mark.parametrize("eps", [0., 1e-2])
-  def test_gw_better_initialization_helps(self, rng: jnp.ndarray, eps: float):
-    n, m, d1, d2, rank = 123, 124, 12, 10, 5
-    key1, key2, key3, key4 = jax.random.split(rng, 4)
-
-    geom_x = pointcloud.PointCloud(
-        jax.random.normal(key1, (n, d1)),
-        jax.random.normal(key2, (n, d1)),
-        epsilon=eps,
-    )
-    geom_y = pointcloud.PointCloud(
-        jax.random.normal(key3, (m, d2)),
-        jax.random.normal(key4, (m, d2)),
-        epsilon=eps,
-    )
-    problem = quadratic_problem.QuadraticProblem(geom_x, geom_y)
-    solver_random = gromov_wasserstein.GromovWasserstein(
-        rank=rank,
-        initializer="random",
-        quad_initializer="random",
-        epsilon=eps,
-        store_inner_errors=True,
-    )
-    solver_kmeans = gromov_wasserstein.GromovWasserstein(
-        rank=rank,
-        initializer="k-means",
-        quad_initializer="k-means",
-        epsilon=eps,
-        store_inner_errors=True
-    )
-
-    out_random = solver_random(problem)
-    out_kmeans = solver_kmeans(problem)
-
-    assert out_random.reg_gw_cost - out_kmeans.reg_gw_cost >= 1.
-    random_errors = out_random.errors[out_random.errors > -1]
-    kmeans_errors = out_kmeans.errors[out_kmeans.errors > -1]
-    np.testing.assert_array_equal(random_errors >= 0., True)
-    np.testing.assert_array_equal(kmeans_errors >= 0., True)
