+++ tests/problems/linear/potentials_test.py
