+++ tests/solvers/linear/sinkhorn_diff_test.py
