+++ tests/solvers/linear/sinkhorn_grid_test.py
