+++ tests/solvers/linear/sinkhorn_lr_test.py
