+++ tests/solvers/linear/sinkhorn_test.py
